fix: put the blank pages components to separate folders and update routes
