+++ client/src/components/pages/result-page/result-page.tsx
