+++ client/src/components/pages/welcome-page/welcome-page.tsx
